requirements.txt
confluent-kafka
requests
six
python-dotenv
pandas
jsonschema==3.2.0
pymongo[srv]==4.1.1
dnspython
ipykernel 

-e .
